PiecewiseVanillaModel

Documentation for PiecewiseVanillaModel.

PiecewiseVanillaModel._d_1

PiecewiseVanillaModel.bachelier_implied_stdev_numeric

PiecewiseVanillaModel.bachelier_implied_volatility_numeric

PiecewiseVanillaModel.bachelier_price

PiecewiseVanillaModel.bachelier_price

PiecewiseVanillaModel.bachelier_vega

PiecewiseVanillaModel.bachelier_vega

PiecewiseVanillaModel.black_price

PiecewiseVanillaModel.black_price

PiecewiseVanillaModel._d_1 — Method

_d_1(strike, forward, nu)

Calculate the Black formula d_1 term.

We implement this once in order to fine-tune corner cases later.

PiecewiseVanillaModel.bachelier_implied_stdev_numeric — Function

bachelier_implied_stdev(price, strike, forward, call_put, min_max = (1.0e-4, 6.0e-2))

Calculate the implied normal standard deviation ν from a Bachelier model price.

PiecewiseVanillaModel.bachelier_implied_volatility_numeric — Function

bachelier_implied_volatility(price, strike, forward, T, call_put, min_max = (0.0001, 0.02))

Calculate the implied normal volatility σ from a Bachelier model price.

PiecewiseVanillaModel.bachelier_price — Method

bachelier_price(strike, forward, nu, call_put)

Calculate Vanilla option price $V$ in Bachelier model.

Argument strike ($K$) represents the option strike, forward ($F$) is the forward price of the underlying $S$, i.e. $F = E[S(T)]$ in the respective pricing measure.

Argument nu represents the standard deviation of the forward price. For annualised normal volatility σ, we have ν = σ √T. Finally, call_put is the call (+1) or put (-1) option flag.

PiecewiseVanillaModel.bachelier_price — Method

bachelier_price(strike, forward, σ, T, call_put)

Calculate Vanilla option price $V$ in Bachelier model with volatility parameter.

PiecewiseVanillaModel.bachelier_vega — Method

bachelier_vega(strike, forward, σ, T)

Calculate Vanilla option Vega in Bachelier model with volatility parameter.

Here, Vega is calculated as $dV / dσ$.

PiecewiseVanillaModel.bachelier_vega — Method

bachelier_vega(strike, forward, nu)

Calculate Vanilla option Vega in Bachelier model.

Here, Vega is calculated as $dV / d ν$.

PiecewiseVanillaModel.black_price — Method

black_price(strike, forward, nu, call_put)

Calculate Vanilla option price V in Black model.

Argument strike ($K$) represents the option strike, forward ($F$) is the forward price of the underlying $S$, i.e. $F = E[S(T)]$ in the respective pricing measure.

Argument nu represents the standard deviation of the forward price. For annualised lognormal volatility σ, we have $ν = σ √T$. Finally, call_put is the call (+1) or put (-1) option flag.

We allow broadcasting for arguments.

PiecewiseVanillaModel.black_price — Method

black_price(strike, forward, σ, T, call_put)

Calculate Vanilla option price $V$ in Black model with volatility parameter.