PiecewiseVanillaModel

Documentation for PiecewiseVanillaModel.

PiecewiseVanillaModel._d_1

PiecewiseVanillaModel.bachelier_implied_stdev_numeric

PiecewiseVanillaModel.bachelier_implied_volatility

PiecewiseVanillaModel.bachelier_implied_volatility_numeric

PiecewiseVanillaModel.bachelier_price

PiecewiseVanillaModel.bachelier_price

PiecewiseVanillaModel.bachelier_vega

PiecewiseVanillaModel.bachelier_vega

PiecewiseVanillaModel.black_implied_volatility

PiecewiseVanillaModel.black_price

PiecewiseVanillaModel.black_price

PiecewiseVanillaModel.brownian_factor

PiecewiseVanillaModel.brownian_factor

PiecewiseVanillaModel.brownian_grid

PiecewiseVanillaModel.calibrated_model

PiecewiseVanillaModel.calibrated_model_from_slopes

PiecewiseVanillaModel.calibrated_model_from_smile

PiecewiseVanillaModel.call_option

PiecewiseVanillaModel.cum_dist

PiecewiseVanillaModel.implied_density

PiecewiseVanillaModel.initial_values

PiecewiseVanillaModel.integral_one

PiecewiseVanillaModel.integral_two

PiecewiseVanillaModel.local_volatility

PiecewiseVanillaModel.local_volatility

PiecewiseVanillaModel.lognormal_volatility

PiecewiseVanillaModel.model

PiecewiseVanillaModel.normal_volatility

PiecewiseVanillaModel.normal_volatility

PiecewiseVanillaModel.positive_volatilities

PiecewiseVanillaModel.power_call_option

PiecewiseVanillaModel.power_put_option

PiecewiseVanillaModel.put_option

PiecewiseVanillaModel.risk_factor

PiecewiseVanillaModel.risk_factor

PiecewiseVanillaModel.vanilla_option_integral

PiecewiseVanillaModel.variance

PiecewiseVanillaModel._d_1 — Method

_d_1(strike, forward, nu)

Calculate the Black formula d_1 term.

We implement this once in order to fine-tune corner cases later.

PiecewiseVanillaModel.bachelier_implied_stdev_numeric — Function

bachelier_implied_stdev(price, strike, forward, call_put, min_max = (1.0e-4, 6.0e-2))

Calculate the implied normal standard deviation ν from a Bachelier model price.

PiecewiseVanillaModel.bachelier_implied_volatility — Method

bachelier_implied_volatility(price, strike, forward, T, call_put)

Calculate the Bachelier or normal implied volatility.

price is the forward or undiscounted option price, strike is the option strike, forward is the forward asset price or expectation of factor, T is time to option expiry, and call_put encodes call (+1) or put (-1) options.

PiecewiseVanillaModel.bachelier_implied_volatility_numeric — Function

bachelier_implied_volatility(price, strike, forward, T, call_put, min_max = (0.0001, 0.02))

Calculate the implied normal volatility σ from a Bachelier model price.

PiecewiseVanillaModel.bachelier_price — Method

bachelier_price(strike, forward, nu, call_put)

Calculate Vanilla option price $V$ in Bachelier model.

Argument strike ($K$) represents the option strike, forward ($F$) is the forward price of the underlying $S$, i.e. $F = E[S(T)]$ in the respective pricing measure.

Argument nu represents the standard deviation of the forward price. For annualised normal volatility σ, we have ν = σ √T. Finally, call_put is the call (+1) or put (-1) option flag.

PiecewiseVanillaModel.bachelier_price — Method

bachelier_price(strike, forward, σ, T, call_put)

Calculate Vanilla option price $V$ in Bachelier model with volatility parameter.

PiecewiseVanillaModel.bachelier_vega — Method

bachelier_vega(strike, forward, σ, T)

Calculate Vanilla option Vega in Bachelier model with volatility parameter.

Here, Vega is calculated as $dV / dσ$.

PiecewiseVanillaModel.bachelier_vega — Method

bachelier_vega(strike, forward, nu)

Calculate Vanilla option Vega in Bachelier model.

Here, Vega is calculated as $dV / d ν$.

PiecewiseVanillaModel.black_implied_volatility — Method

black_implied_volatility(price, strike, forward, T, call_put)

Calculate the Black or log-normal implied volatility.

price is the forward or undiscounted option price, strike is the option strike, forward is the forward asset price or expectation of factor, T is time to option expiry, and call_put encodes call (+1) or put (-1) options.

PiecewiseVanillaModel.black_price — Method

black_price(strike, forward, nu, call_put)

Calculate Vanilla option price V in Black model.

Argument strike ($K$) represents the option strike, forward ($F$) is the forward price of the underlying $S$, i.e. $F = E[S(T)]$ in the respective pricing measure.

Argument nu represents the standard deviation of the forward price. For annualised lognormal volatility σ, we have $ν = σ √T$. Finally, call_put is the call (+1) or put (-1) option flag.

We allow broadcasting for arguments.

PiecewiseVanillaModel.black_price — Method

black_price(strike, forward, σ, T, call_put)

Calculate Vanilla option price $V$ in Black model with volatility parameter.

PiecewiseVanillaModel.brownian_factor — Method

brownian_factor(ds_, v0, ds, dv, dw, r_extrapolation)

Calculate the auxiliary parameter w from a given risk factor S.

This is an internal method. Risk factor and resulting auxiliary parameter are represented relative to ATM.

PiecewiseVanillaModel.brownian_factor — Method

brownian_factor(m::Model, s)

Calculate the auxiliary parameter w from a given risk factor S.

PiecewiseVanillaModel.brownian_grid — Method

brownian_grid(v0, ds, dv)

Calculate the grid of auxiliary parameters at model creation.

PiecewiseVanillaModel.calibrated_model — Method

calibrated_model(
    s_atm,
    σ_atm,
    T,
    dsl,
    dsu,
    dvl,
    dvu;
    rexl  = 0.0,
    rexu  = 0.0,
    σ_tol = 1.0e-10,
    k_max = 10,
    l_max = 5,
    backtracking_factor = 0.5,
    )

Create a model which is auto-calibrated to at-the-money forward and normal volatility.

PiecewiseVanillaModel.calibrated_model_from_slopes — Method

calibrated_model_from_slopes(
    s_atm,
    σ_atm,
    T,
    std_devs_lo,
    std_devs_up,
    slope_lo,
    slope_up,
    rexl,
    rexu,
    )

Create an ATM-calibrated model from relative/normalised reference strike and volatility slope increments.

This method is used internally for smile calibration.

PiecewiseVanillaModel.calibrated_model_from_smile — Method

calibrated_model_from_smile(
    s_atm,
    σ_atm,
    T,
    std_devs_lo,
    std_devs_up,
    rel_strikes,
    σ_smiles;
    rexl  = 0.0,
    rexu  = 0.0,
    α = 0.0,
    σ_min = 1.0e-4,
    lmfit_kwargs = (
        autodiff = :forwarddiff,
        maxIter  = 10
    ),
    )

Create a model that is calibrated to a set of strikes and implied normal volatilities.

Input implied normal volatilities are σ_smiles. Corresponding strikes are rel_strikes. Strikes are represented as off-set to at-the-money (ATM) forward s_atm.

ATM implied normal volatility is σ_atm. Time to option expiry is T.

Lower and upper reference strikes are represented by std_devs_lo and std_devs_up. The reference strikes are expressed in terms of standard deviations from ATM. Transformation into actual strikes is implemented via the scaling σ_atm*sqrt(T).

Lower and upper smile extrapolation is controlled via the slope parameters rexlo and rexu.

The calibration method is controlled via the named tuple lmfit_kwargs. The arguments are passed on to the LsqFit.lmfit(...) method.

PiecewiseVanillaModel.call_option — Method

call_option(m::Model, strike)

Calculate a call option for a given (out-of-the-money) strike.

PiecewiseVanillaModel.cum_dist — Method

cum_dist(m::Model, s)

Calculate the implied cumulative distribution function for a risk factor level s.

PiecewiseVanillaModel.implied_density — Method

implied_density(m::Model, s)

Calculate model-implied probability density for a given risk factor value.

PiecewiseVanillaModel.initial_values — Method

initial_values(idx, v0, ds, dv, dw, r_extrapolation)

Derive the model parameters for a segment specified by the index idx.

PiecewiseVanillaModel.integral_one — Method

integral_one(w, s0, v0, w0, T, r)

Calculate anti-derivative for S(w).

This is an internal method used vor call and put option calculation.

PiecewiseVanillaModel.integral_two — Method

integral_two(w, s0, v0, w0, T, r)

Calculate anti-derivative for S(w)^2.

This is an internal method used vor power option calculation.

PiecewiseVanillaModel.local_volatility — Method

local_volatility(ds_, v0, ds, dv, dw, r_extrapolation)

Calculate interpolated volatility parameters.

PiecewiseVanillaModel.local_volatility — Method

local_volatility(m::Model, s)

Calculate interpolated volatility parameters.

PiecewiseVanillaModel.lognormal_volatility — Method

lognormal_volatility(m::Model, strike)

Calculate the Black or log-normal implied volatility for a given model.

PiecewiseVanillaModel.model — Method

model(s0, v0, w0, T, dsl, dsu, dvl, dvu; rexl = 0.0, rexu = 0.0)

Create a model from direct inputs and without calibration.

Model attributes are as follows:

s0 is the forward risk factor (at-the-money level).

v0 is the volatility parameter at s0.

w0 is the auxiliary parameter at s0.

T is the time to option expiry.

dsl is a vector of relative reference strikes for lower smile; Actual strikes are sl = s0 - dsl.

dsu is a vector of relative reference strikes for upper smile; Actual strikes are su = s0 + dsu.

dvl is a vector of volatility parameter offsets for lower smile; Actual volatility is vl = v0 + dvl.

dvu is a vector of volatility parameter offsets for upper smile; Actual volatility is vu = v0 + dvu.

rexl is the outward slope of the lower extrapolation smile; log-normal model uses rexl < 0. If rexl == nothing then slope is calculated via linear extrapolation.

rexu is the outward slope of the upper extrapolation smile; log-normal model uses rexl > 0. If rexu == nothing then slope is calculated via linear extrapolation.

PiecewiseVanillaModel.normal_volatility — Method

normal_volatility(b76_volatility, strike, forward, T)

Calculate Bachelier or normal implied volatility from Black or log-normal volatility

PiecewiseVanillaModel.normal_volatility — Method

normal_volatility(m::Model, strike)

Calculate the Bachelier or normal implied volatility for a given model.

PiecewiseVanillaModel.positive_volatilities — Method

positive_volatilities(v0, dv)

Check for resulting positive volatilities and floor volatilities if necessary.

PiecewiseVanillaModel.power_call_option — Method

power_call_option(m::Model, strike)

Calculate a power call option for a given (out-of-the-money) strike.

PiecewiseVanillaModel.power_put_option — Method

power_put_option(m::Model, strike)

Calculate a power put option for a given (out-of-the-money) strike.

PiecewiseVanillaModel.put_option — Method

put_option(m::Model, strike)

Calculate a put option for a given (out-of-the-money) strike.

PiecewiseVanillaModel.risk_factor — Method

risk_factor(dw_, v0, ds, dv, dw, r_extrapolation)

Calculate the risk factor S from a given auxiliary parameter w.

This is an internal method. Auxiliary parameter and resulting risk factor are represented relative to ATM.

PiecewiseVanillaModel.risk_factor — Method

risk_factor(m::Model, w)

Calculate the risk factor S from a given auxiliary parameter w.

PiecewiseVanillaModel.vanilla_option_integral — Method

vanilla_option_integral(m::Model, strike, payoff, call_or_put)

Calculate the option price for a generic payoff function using numerical integration via quadgk(...).

PiecewiseVanillaModel.variance — Method

variance(m::Model)

Calculate the model-implied variance of the risk factor.